moved pages into tabs (preferences into account and invitations into groups) and made some enhancemnents
