+++ frontend/src/app/account/page.jsx
@@ -1,37 +1,73 @@
 'use client';
 
-import { useState, useEffect } from 'react';
-import { 
-  Container, 
-  Title, 
-  Text, 
-  Stack, 
-  Box, 
-  TextInput, 
-  Textarea, 
+import { useState, useEffect, Suspense } from 'react';
+import { useSearchParams } from 'next/navigation';
+import {
+  Container,
+  Title,
+  Text,
+  Stack,
+  Box,
+  TextInput,
+  Textarea,
   Select,
-  Button, 
+  Button,
   Card,
   Group,
   Avatar,
   Loader,
   Alert,
   Divider,
-  Badge
+  Badge,
+  Tabs,
 } from '@mantine/core';
-import { 
-  IconUser, 
-  IconBuilding, 
-  IconSchool, 
-  IconBriefcase, 
-  IconCheck, 
+import {
+  IconUser,
+  IconBuilding,
+  IconSchool,
+  IconBriefcase,
+  IconCheck,
   IconAlertCircle,
   IconMail,
-  IconShield
+  IconShield,
+  IconSettings,
 } from '@tabler/icons-react';
 import { Navigation } from '../components/Navigation';
 import { ProtectedRoute } from '../components/ProtectedRoute';
+import { PreferencesForm } from '../components/PreferencesForm';
 import { useAuth } from '../contexts/AuthContext';
+
+function AccountTabs() {
+  const searchParams = useSearchParams();
+  const tabFromUrl = searchParams.get('tab');
+  const [activeTab, setActiveTab] = useState(tabFromUrl === 'preferences' ? 'preferences' : 'profile');
+
+  useEffect(() => {
+    const t = searchParams.get('tab');
+    if (t === 'preferences') setActiveTab('preferences');
+  }, [searchParams]);
+
+  return (
+    <Tabs value={activeTab} onChange={setActiveTab}>
+      <Tabs.List mb="xl">
+        <Tabs.Tab value="profile" leftSection={<IconUser size={16} />}>
+          Profile
+        </Tabs.Tab>
+        <Tabs.Tab value="preferences" leftSection={<IconSettings size={16} />}>
+          Preferences
+        </Tabs.Tab>
+      </Tabs.List>
+
+      <Tabs.Panel value="profile">
+        <ProfilePanel />
+      </Tabs.Panel>
+
+      <Tabs.Panel value="preferences">
+        <PreferencesForm />
+      </Tabs.Panel>
+    </Tabs>
+  );
+}
 
 export default function AccountPage() {
   return (
@@ -42,8 +78,28 @@ export default function AccountPage() {
 }
 
 function AccountPageContent() {
+  return (
+    <Box style={{ minHeight: '100vh', backgroundColor: '#fafafa' }}>
+      <Navigation />
+
+      <Container size="lg" py="xl">
+        <Stack gap="xl">
+          <Stack align="center" gap="md">
+            <Title order={1}>Your Account</Title>
+          </Stack>
+
+          <Suspense fallback={<Loader />}>
+            <AccountTabs />
+          </Suspense>
+        </Stack>
+      </Container>
+    </Box>
+  );
+}
+
+function ProfilePanel() {
   const { authState, getValidToken } = useAuth();
-  
+
   const [loading, setLoading] = useState(true);
   const [saving, setSaving] = useState(false);
   const [error, setError] = useState(null);
@@ -51,8 +107,7 @@ function AccountPageContent() {
   const [companyOptions, setCompanyOptions] = useState([]);
   const [schoolOptions, setSchoolOptions] = useState([]);
   const [roleTitleOptions, setRoleTitleOptions] = useState([]);
-  
-  // User profile data
+
   const [userData, setUserData] = useState({
     id: '',
     email: '',
@@ -62,26 +117,24 @@ function AccountPageContent() {
     school_name: '',
     role_title: '',
     verification_status: 'unverified',
-    profile_picture_url: ''
+    profile_picture_url: '',
   });
 
-  // Fetch user data on mount
   useEffect(() => {
     const fetchUserData = async () => {
       if (!authState?.accessToken) return;
-      
+
       try {
         const token = await getValidToken();
         const response = await fetch('http://localhost:8000/api/auth/me', {
           headers: {
-            'Authorization': `Bearer ${token}`
-          }
+            'Authorization': `Bearer ${token}`,
+          },
         });
-        
+
         const data = await response.json();
-        
+
         if (response.ok && data.user) {
-          // Profile data is in data.user.profile, email is in data.user.email
           const profile = data.user.profile || {};
           setUserData({
             id: profile.id ?? '',
@@ -92,7 +145,7 @@ function AccountPageContent() {
             school_name: profile.school_name ?? '',
             role_title: profile.role_title ?? '',
             verification_status: profile.verification_status ?? 'unverified',
-            profile_picture_url: profile.profile_picture_url ?? ''
+            profile_picture_url: profile.profile_picture_url ?? '',
           });
         } else {
           setError('Failed to load account data');
@@ -104,7 +157,7 @@ function AccountPageContent() {
         setLoading(false);
       }
     };
-    
+
     fetchUserData();
   }, [authState, getValidToken]);
 
@@ -137,63 +190,58 @@ function AccountPageContent() {
     loadOptions();
   }, []);
 
-  // Handle form field changes
   const handleChange = (field, value) => {
-    setUserData(prev => ({
+    setUserData((prev) => ({
       ...prev,
-      [field]: value
+      [field]: value,
     }));
-    // Clear messages when user starts editing
     setError(null);
     setSuccess(null);
   };
 
-  // Save changes
   const handleSave = async () => {
     if (!userData.id) {
       setError('User ID not found');
       return;
     }
-    
+
     setSaving(true);
     setError(null);
     setSuccess(null);
-    
+
     try {
       const token = await getValidToken();
-      
-      // Prepare update data (only send editable fields)
+
       const updateData = {
         full_name: userData.full_name,
         bio: userData.bio || null,
         company_name: userData.company_name || null,
         school_name: userData.school_name || null,
-        role_title: userData.role_title || null
+        role_title: userData.role_title || null,
       };
-      
+
       const response = await fetch(`http://localhost:8000/api/users/${userData.id}`, {
         method: 'PUT',
         headers: {
           'Content-Type': 'application/json',
-          'Authorization': `Bearer ${token}`
+          'Authorization': `Bearer ${token}`,
         },
-        body: JSON.stringify(updateData)
+        body: JSON.stringify(updateData),
       });
-      
+
       const data = await response.json();
-      
+
       if (response.ok && data.status === 'success') {
         setSuccess('Profile updated successfully!');
-        // Update local state with response data (ensure no nulls)
         if (data.data) {
           const updated = data.data;
-          setUserData(prev => ({
+          setUserData((prev) => ({
             ...prev,
             full_name: updated.full_name ?? prev.full_name,
             bio: updated.bio ?? '',
             company_name: updated.company_name ?? '',
             school_name: updated.school_name ?? '',
-            role_title: updated.role_title ?? ''
+            role_title: updated.role_title ?? '',
           }));
         }
       } else {
@@ -220,137 +268,116 @@ function AccountPageContent() {
 
   if (loading) {
     return (
-      <Box style={{ minHeight: '100vh', backgroundColor: '#fafafa' }}>
-        <Navigation />
-        <Container size="md" py="xl">
-          <Stack align="center" gap="md" style={{ minHeight: '400px', justifyContent: 'center' }}>
-            <Loader size="lg" />
-            <Text>Loading your account...</Text>
-          </Stack>
-        </Container>
-      </Box>
+      <Stack align="center" gap="md" py="xl" style={{ minHeight: '300px', justifyContent: 'center' }}>
+        <Loader size="lg" />
+        <Text>Loading your profile...</Text>
+      </Stack>
     );
   }
 
   return (
-    <Box style={{ minHeight: '100vh', backgroundColor: '#fafafa' }}>
-      <Navigation />
-      
-      <Container size="md" py="xl">
-        <Stack gap="xl">
-          {/* Header */}
-          <Stack align="center" gap="md">
-            <Avatar 
-              src={userData.profile_picture_url} 
-              size={100} 
-              radius="50%"
-              color="blue"
-            >
-              {userData.full_name?.charAt(0)?.toUpperCase() || 'U'}
-            </Avatar>
-            <Title order={1}>Your Account</Title>
-            <Group gap="xs">
-              {getVerificationBadge()}
-            </Group>
-          </Stack>
+    <Stack gap="xl">
+      {/* Avatar & verification */}
+      <Stack align="center" gap="md">
+        <Avatar
+          src={userData.profile_picture_url}
+          size={100}
+          radius="50%"
+          color="blue"
+        >
+          {userData.full_name?.charAt(0)?.toUpperCase() || 'U'}
+        </Avatar>
+        <Group gap="xs">
+          {getVerificationBadge()}
+        </Group>
+      </Stack>
 
-          {/* Messages */}
-          {error && (
-            <Alert icon={<IconAlertCircle size={16} />} color="red" title="Error">
-              {error}
-            </Alert>
-          )}
-          
-          {success && (
-            <Alert icon={<IconCheck size={16} />} color="green" title="Success">
-              {success}
-            </Alert>
-          )}
+      {error && (
+        <Alert icon={<IconAlertCircle size={16} />} color="red" title="Error">
+          {error}
+        </Alert>
+      )}
 
-          {/* Profile Card */}
-          <Card withBorder padding="xl" radius="md">
-            <Stack gap="lg">
-              <Title order={3}>Profile Information</Title>
-              
-              {/* Email (read-only) */}
-              <TextInput
-                label="Email"
-                value={userData.email}
-                disabled
-                leftSection={<IconMail size={16} />}
-                description="Email cannot be changed"
-              />
-              
-              {/* Full Name */}
-              <TextInput
-                label="Full Name"
-                placeholder="Your full name"
-                value={userData.full_name}
-                onChange={(e) => handleChange('full_name', e.target.value)}
-                leftSection={<IconUser size={16} />}
-                required
-              />
-              
-              {/* Bio */}
-              <Textarea
-                label="Bio"
-                placeholder="Tell others about yourself..."
-                value={userData.bio}
-                onChange={(e) => handleChange('bio', e.target.value)}
-                minRows={3}
-                maxRows={5}
-              />
-              
-              <Divider label="Work & Education" labelPosition="center" />
-              
-              {/* Company */}
-              <Select
-                label="Company"
-                placeholder="Select company"
-                data={companyOptions}
-                searchable
-                value={userData.company_name}
-                onChange={(value) => handleChange('company_name', value || '')}
-                leftSection={<IconBuilding size={16} />}
-              />
-              
-              {/* School */}
-              <Select
-                label="School"
-                placeholder="Select school"
-                data={schoolOptions}
-                searchable
-                value={userData.school_name}
-                onChange={(value) => handleChange('school_name', value || '')}
-                leftSection={<IconSchool size={16} />}
-              />
-              
-              {/* Role/Title */}
-              <Select
-                label="Job Title / Role"
-                placeholder="Select role/title"
-                data={roleTitleOptions}
-                searchable
-                value={userData.role_title}
-                onChange={(value) => handleChange('role_title', value || '')}
-                leftSection={<IconBriefcase size={16} />}
-              />
-            </Stack>
-          </Card>
+      {success && (
+        <Alert icon={<IconCheck size={16} />} color="green" title="Success">
+          {success}
+        </Alert>
+      )}
 
-          {/* Save Button */}
-          <Group justify="center">
-            <Button 
-              size="lg" 
-              onClick={handleSave}
-              loading={saving}
-              leftSection={<IconCheck size={18} />}
-            >
-              Save Changes
-            </Button>
-          </Group>
+      <Card withBorder padding="xl" radius="md">
+        <Stack gap="lg">
+          <Title order={3}>Profile Information</Title>
+
+          <TextInput
+            label="Email"
+            value={userData.email}
+            disabled
+            leftSection={<IconMail size={16} />}
+            description="Email cannot be changed"
+          />
+
+          <TextInput
+            label="Full Name"
+            placeholder="Your full name"
+            value={userData.full_name}
+            onChange={(e) => handleChange('full_name', e.target.value)}
+            leftSection={<IconUser size={16} />}
+            required
+          />
+
+          <Textarea
+            label="Bio"
+            placeholder="Tell others about yourself..."
+            value={userData.bio}
+            onChange={(e) => handleChange('bio', e.target.value)}
+            minRows={3}
+            maxRows={5}
+          />
+
+          <Divider label="Work & Education" labelPosition="center" />
+
+          <Select
+            label="Company"
+            placeholder="Select company"
+            data={companyOptions}
+            searchable
+            value={userData.company_name}
+            onChange={(value) => handleChange('company_name', value || '')}
+            leftSection={<IconBuilding size={16} />}
+          />
+
+          <Select
+            label="School"
+            placeholder="Select school"
+            data={schoolOptions}
+            searchable
+            value={userData.school_name}
+            onChange={(value) => handleChange('school_name', value || '')}
+            leftSection={<IconSchool size={16} />}
+          />
+
+          <Select
+            label="Job Title / Role"
+            placeholder="Select role/title"
+            data={roleTitleOptions}
+            searchable
+            value={userData.role_title}
+            onChange={(value) => handleChange('role_title', value || '')}
+            leftSection={<IconBriefcase size={16} />}
+          />
         </Stack>
-      </Container>
-    </Box>
+      </Card>
+
+      <Group justify="center">
+        <Button
+          size="lg"
+          onClick={handleSave}
+          loading={saving}
+          leftSection={<IconCheck size={18} />}
+        >
+          Save Changes
+        </Button>
+      </Group>
+    </Stack>
   );
 }
