+++ frontend/src/app/components/Navigation.jsx
@@ -52,20 +52,6 @@ export function Navigation() {
                       </Text>
                     </UnstyledButton>
                   </Link>
-                  <Link href="/invitations" style={{ textDecoration: 'none' }} data-tour="nav-invitations">
-                    <UnstyledButton>
-                      <Text size="md" c="#666" style={{ transition: 'color 0.2s' }}>
-                        Invitations
-                      </Text>
-                    </UnstyledButton>
-                  </Link>
-                  <Link href="/preferences" style={{ textDecoration: 'none' }} data-tour="nav-preferences">
-                    <UnstyledButton>
-                      <Text size="md" c="#666" style={{ transition: 'color 0.2s' }}>
-                        Preferences
-                      </Text>
-                    </UnstyledButton>
-                  </Link>
                   <Link href="/discover" style={{ textDecoration: 'none' }} data-tour="nav-discover">
                     <UnstyledButton>
                       <Text size="md" c="#666" style={{ transition: 'color 0.2s' }}>
@@ -168,20 +154,6 @@ export function Navigation() {
                   <UnstyledButton style={{ width: '100%' }}>
                     <Text size="lg" c="#666" style={{ padding: '0.5rem 0' }}>
                       Groups
-                    </Text>
-                  </UnstyledButton>
-                </Link>
-                <Link href="/invitations" style={{ textDecoration: 'none' }} onClick={handleLinkClick}>
-                  <UnstyledButton style={{ width: '100%' }}>
-                    <Text size="lg" c="#666" style={{ padding: '0.5rem 0' }}>
-                      Invitations
-                    </Text>
-                  </UnstyledButton>
-                </Link>
-                <Link href="/preferences" style={{ textDecoration: 'none' }} onClick={handleLinkClick}>
-                  <UnstyledButton style={{ width: '100%' }}>
-                    <Text size="lg" c="#666" style={{ padding: '0.5rem 0' }}>
-                      Preferences
                     </Text>
                   </UnstyledButton>
                 </Link>
